requirements.txt
aniso8601==7.0.0
asgiref==3.6.0
Django==3.2.19
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
PyJWT==1.7.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
django-cors-headers==4.0.0
psycopg2==2.9.6
django-graphql-auth==0.3.16
sentry-sdk
django-cors-middleware
Pillow==9.5.0
graphene-file-upload==1.3.0
djangorestframework~=3.14.0
stripe==5.5.0
django-storages==1.13.2
boto3==1.28.5
channels==4.0.0
channels-redis==4.1.0
django-recaptcha==3.0.0
django-allauth==0.58.1